refactor(api): convert the remaining params
There remain a handful of raw accessed to `params`, however these
require more complex validation that is provided via the accessors in
their current form.

The use of the `Params` object is good.
Ultimately we should use the params object for everything, and to achive
it, we can use modules of param definitions (inlcluding validation).

This would centralise the validations, and ensure minimal duplication of
params across the API surface.